% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticateUser.R
\name{authenticateUser}
\alias{authenticateUser}
\title{Authenticate API}
\usage{
authenticateUser()
}
\value{
url
}
\description{
Authenticate API to get access to data
}
\examples{
authenticateUser()

}
