% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenomicsResponse.R
\name{convert_list_to_dataframe}
\alias{convert_list_to_dataframe}
\title{Get response from genomics endpoint}
\usage{
convert_list_to_dataframe(list_obj)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
