% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemiologyDataDictionary.R
\name{epidemiologyDataDictionary}
\alias{epidemiologyDataDictionary}
\title{Documentation of Epidemiology API fields}
\usage{
epidemiologyDataDictionary()
}
\value{
dataframe
}
\description{
Documents fields included when extracting data from outbreak.info
}
\examples{
df = epidemiologyDataDictionary()
knitr::kable(df)

}
