% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicsDataDictionary.R
\name{genomicsDataDictionary}
\alias{genomicsDataDictionary}
\title{Documentation of genomics (SARS-CoV-2 variant prevalence) API fields}
\usage{
genomicsDataDictionary()
}
\value{
dataframe
}
\description{
Returns a dataframe describing the fields included when making calls to the genomics API.
}
\examples{
knitr::kable(genomicsDataDictionary())
}
