% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmn0.R
\name{getAdmn0}
\alias{getAdmn0}
\title{Retrieve COVID-19 data for all countries}
\usage{
getAdmn0(...)
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for all countries
}
\examples{
\dontrun{
getAdmn0()
}
}
