% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmn1ByCountry.R
\name{getAdmn1ByCountry}
\alias{getAdmn1ByCountry}
\title{Retrieve COVID-19 data for states/provinces in a country or countries}
\usage{
getAdmn1ByCountry(countries, ...)
}
\arguments{
\item{countries}{vector or list of country name(s)}
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for all states/provinces in one or more countries.
}
\examples{
getAdmn1ByCountry("India")

}
