% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmn2ByCountry.R
\name{getAdmn2ByCountry}
\alias{getAdmn2ByCountry}
\title{Retrieve COVID-19 data for all counties in the US.}
\usage{
getAdmn2ByCountry(...)
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for all counties in the United States of America.
}
\examples{
# Takes a long time to run!
\dontrun{
getAdmn2ByCountry()
}

}
