% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllLineagesByLocation.R
\name{getAllLineagesByLocation}
\alias{getAllLineagesByLocation}
\title{Retrieve prevalence of lineages by location}
\usage{
getAllLineagesByLocation(
  location,
  other_threshold = 0.05,
  nday_threshold = 10,
  ndays = 180,
  other_exclude = NULL,
  cumulative = F
)
}
\arguments{
\item{location}{a location name}

\item{other_threshold}{minimum prevalence threshold below which lineages are accumulated under "Other" (default=0.05)}

\item{nday_threshold}{minimum number of days in which the prevalence of a lineage must be below other_threshold to be accumulated under "Other" (default=10)}

\item{ndays}{the number of days before the current date to be used as a window to accumulate lineages under "Other" (default=180)}

\item{other_exclude}{(optional) lineage(s) that are NOT to be included under "Other" even if the conditions specified by the three thresholds above are met}

\item{cumulative}{`Boolean` (T/F), T returns cumulative prevalence of lineages (default=F)}
}
\value{
dataframe
}
\description{
Retrieve prevalence of all lineages above a specified frequency over time by location
}
\examples{
india = getAllLineagesByLocation(location = "India", other_threshold=0.03, ndays=60)
india[india$date == "2021-06-25",]

}
