% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getByAdmnLevel.R
\name{getByAdmnLevel}
\alias{getByAdmnLevel}
\title{Retrieve COVID-19 data for all locations at an administrative level.}
\usage{
getByAdmnLevel(admin_level, ...)
}
\arguments{
\item{admin_level}{an integer representing an administrative level (World Bank regions = -1, countries = 0, states/provinces = 1, metropolitan areas = 1.5, counties = 2)}
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for all locations at a given administrative level.
}
\examples{
getByAdmnLevel(-1)

}
