% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountryByRegion.R
\name{getCountryByRegion}
\alias{getCountryByRegion}
\title{Retrieve COVID-19 data for countries in a World Bank region(s)}
\usage{
getCountryByRegion(wb_regions, ...)
}
\arguments{
\item{wb_regions}{vector or list of World Bank region(s)}
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for all countries in one or more World Bank region(s).
}
\examples{
getCountryByRegion("North America")

}
