% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCumulativeBySubadmin.R
\name{getCumulativeBySubadmin}
\alias{getCumulativeBySubadmin}
\title{Retrieve cumulative prevalence of a lineage}
\usage{
getCumulativeBySubadmin(
  pangolin_lineage,
  location = NULL,
  mutations = NULL,
  ndays = NULL
)
}
\arguments{
\item{pangolin_lineage}{PANGO lineage name}

\item{location}{(optional) a location name (if not specified, cumulative prevalence at the country level globally returned)}

\item{mutations}{(optional) a `vector` of mutation(s)}

\item{ndays}{(optional) an `integer` specifying number of days from current date to calculative cumuative counts (if not specified, no limit on window)}
}
\value{
dataframe
}
\description{
Retrieve cumulative prevalence of a PANGO lineage by the immediate administrative level of a location
}
\examples{
# Worldwide prevalence of B.1.1.7 by country
head(getCumulativeBySubadmin(pangolin_lineage="B.1.1.7"))

# County-level prevalence of B.1.1.7
head(getCumulativeBySubadmin(pangolin_lineage="B.1.1.7", location="California"))


}
