% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCuratedLineages.R
\name{getCuratedLineages}
\alias{getCuratedLineages}
\title{Lookup curated lienages from outbreak.info}
\usage{
getCuratedLineages()
}
\value{
dataframe containing VOCs, VOIs, VUMs, and associated metadata.
}
\description{
Retrieve a curated list of Variants of Concern, Variants of Interest,
Variants Under Monitoring, and de-escalated variants maintained by the [outbreak.info team](https://outbreak.info/situation-reports)
}
\examples{
curated = getCuratedLineages()
# Pull out the curated lineages which are WHO-desginated
curated \%>\% filter(!is.na(who_name)) \%>\% knitr::kable()
# Pull out the Variants of Concern
curated \%>\% filter(variantType == "Variant of Concern") \%>\% knitr::kable()
}
