% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpiData.R
\name{getEpiData}
\alias{getEpiData}
\title{Retrieve epidemiological data from outbreak.info}
\usage{
getEpiData(
  name = NULL,
  location_id = NULL,
  wb_region = NULL,
  country_name = NULL,
  state_name = NULL,
  admin_level = NULL,
  date = NULL,
  mostRecent = NULL,
  fields = NULL,
  sort = NULL,
  size = 1000
)
}
\arguments{
\item{name}{vector of location names}

\item{location_id}{vector of ISO3 code representing locations}

\item{wb_region}{World Bank region name(s)}

\item{country_name}{country name(s)}

\item{state_name}{state name(s)}

\item{admin_level}{an integer representing an administrative level (World Bank regions = -1, countries = 0, states/provinces = 1, metropolitan areas = 1.5, counties = 2)}

\item{date}{date(s) (YYYY-MM-DD)}

\item{mostRecent}{T/F}

\item{fields}{vector of API fields to include in results}

\item{sort}{parameter to sort results by}

\item{size}{size}
}
\value{
dataframe
}
\description{
Retrieve up-to-date epidemiological data from outbreak.info according to user specifications
}
\examples{
getEpiData(name="United States of America", date="2020-07-01")

}
