% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenomicData.R
\name{getGenomicData}
\alias{getGenomicData}
\title{Retrieve data from api.outbreak.info/genomics}
\usage{
getGenomicData(
  query_url,
  location = NULL,
  cumulative = NULL,
  pangolin_lineage = NULL,
  mutations = NULL,
  ndays = NULL,
  frequency = NULL,
  subadmin = NULL,
  other_threshold = NULL,
  nday_threshold = NULL,
  other_exclude = NULL,
  logInfo = TRUE
)
}
\value{
dataframe
}
\description{
Retrieves up-to-date COVID-19 genomic data from outbreak.info according to user specifications.
}
\examples{
uk_b117 = getGenomicData(query_url="prevalence-by-location", location="United Kingdom", pangolin_lineage = "B.1.1.7")
head(uk_b117)
}
