% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenomicsResponse.R
\name{getGenomicsResponse}
\alias{getGenomicsResponse}
\title{Get response from genomics endpoint}
\usage{
getGenomicsResponse(dataurl, logInfo = T, logWarning = T, logError = T)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
