% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getISO3.R
\name{getISO3}
\alias{getISO3}
\title{Get ISO3 codes}
\usage{
getISO3(locations_to_search)
}
\arguments{
\item{locations_to_search}{vector or list of location (World Bank region, country, state/province, metropolitan area, county) name(s)}
}
\value{
a vector or list of ISO3 codes
}
\description{
Get ISO3 codes for World Bank regions, countries, states/provinces, metropolitan areas, and/or counties.
}
\examples{
getISO3(c("San Diego", "Virginia", "India"))

}
