% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getISO3_genomic.R
\name{getISO3_genomic}
\alias{getISO3_genomic}
\title{Get ISO3 codes from genomic API}
\usage{
getISO3_genomic(locations_to_search)
}
\arguments{
\item{locations_to_search:}{a location name}
}
\value{
an ISO3 code
}
\description{
Get ISO3 codes for World Bank regions, countries, states/provinces, metropolitan areas, and/or counties from genomic API
}
\examples{
getISO3_genomic("San Diego")

}
