% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationData.R
\name{getLocationData}
\alias{getLocationData}
\title{Retrieve COVID-19 data for given locations}
\usage{
getLocationData(location_names, ...)
}
\arguments{
\item{location_names}{vector or list of location name(s)}
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for specified location(s)
}
\examples{
getLocationData(c("California", "India"))

}
