% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationIdGenomic.R
\name{getLocationIdGenomic}
\alias{getLocationIdGenomic}
\title{Get location id codes from genomic API}
\usage{
getLocationIdGenomic(locations_to_search)
}
\arguments{
\item{locations_to_search}{a location name}
}
\value{
a location ID code (ISO3, FIPS)
}
\description{
Get location ID code for countries, states/provinces, metropolitan areas, and/or counties from genomic API
}
\examples{
getLocationIdGenomic("San Diego")

}
