% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetroByCountry.R
\name{getMetroByCountry}
\alias{getMetroByCountry}
\title{Retrieve COVID-19 data for all metropolitan areas in the US.}
\usage{
getMetroByCountry(...)
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for all metropolitan areas in the United States of America.
}
\examples{
\dontrun{
getMetroByCountry()
}
}
