% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutationAcrossLineage.R
\name{getMutationAcrossLineage}
\alias{getMutationAcrossLineage}
\title{Retrieve prevalence of mutation across lineages by location}
\usage{
getMutationAcrossLineage(mutations, location = NULL)
}
\arguments{
\item{mutations}{a `vector` of mutation(s)}

\item{location}{(optional) a location name (if not specified, prevalence globally returned)}
}
\value{
dataframe
}
\description{
Retrieve prevalence of a specified mutation across all lineages by location
}
\examples{
getMutationAcrossLineage(mutations="S:N501Y", location="United States")

}
