% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutationDetails.R
\name{getMutationDetails}
\alias{getMutationDetails}
\title{Retrieve mutation details}
\usage{
getMutationDetails(mutations)
}
\arguments{
\item{mutations}{a `vector` of mutation(s)}
}
\value{
dataframe
}
\description{
Retrieve details of specified mutation(s)
}
\examples{
getMutationDetails(mutations=c("S:E484K", "S:N501Y"))

}
