% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutationsByLineage.R
\name{getMutationsByLineage}
\alias{getMutationsByLineage}
\title{Retrieve mutations by lineage}
\usage{
getMutationsByLineage(pangolin_lineage, frequency = 0.8)
}
\arguments{
\item{pangolin_lineage:}{PANGO lineage name or list}

\item{frequency:}{a number between 0 and 1 specifying the frequency threshold above which to return mutations (default=0.8)}
}
\value{
dataframe
}
\description{
Retrieve all mutations in a specified lineage above a threshold
}
\examples{
getMutationsByLineage(pangolin_lineage="P.1", frequency=0.8)

# get all delta lineages
getMutationsByLineage(pangolin_lineage="P.1", frequency=0.8)
}
