% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalenceByLocation.R
\name{getPrevalenceByLocation}
\alias{getPrevalenceByLocation}
\title{Retrieve daily prevalence of a lineage by location}
\usage{
getPrevalenceByLocation(
  pangolin_lineage,
  location,
  mutations = NULL,
  cumulative = NULL
)
}
\arguments{
\item{pangolin_lineage:}{PANGO lineage name or list of PANGO lineage names. A list of lineages will return a long dataframe with}

\item{location:}{a location name}

\item{mutations:}{(optional) a `vector` of mutation(s)}

\item{cumulative:}{(optional) `Boolean` (T/F), T returns cumulative prevalence since first day of detection}
}
\value{
dataframe
}
\description{
Retrieves the daily prevalence of a PANGO lineage(s) by location
}
\examples{
# P.1 in Brazil
getPrevalenceByLocation(pangolin_lineage = "P.1", location = "Brazil") \%>\% head()

# AY.4 and B.1.617.2 in Brazil (AY.4 and B.1.617.2 separately)
getPrevalenceByLocation(pangolin_lineage = c("AY.4", "B.1.617.2"), location = "Brazil") \%>\% filter(date == "2021-08-01")

# AY.4 OR B.1.617.2 in Brazil (AY.4 and B.1.617.2 combined)
getPrevalenceByLocation(pangolin_lineage = "AY.4 OR B.1.617.2", location = "Brazil")\%>\% filter(date == "2021-08-01")

}
