% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResourcesData.R
\name{getResourcesData}
\alias{getResourcesData}
\title{Title}
\usage{
getResourcesData(
  query = NULL,
  types = NULL,
  size = 10,
  fetchAll = FALSE,
  fields = null,
  sort = NULL,
  facets = NULL,
  facet_size = 10
)
}
\arguments{
\item{query}{(optional): constructs a query over ALL fields, or a fielded query searching within specific fields. Fielded query terms should be separated by ` AND ` or ` OR `. See Elasticserach query strings for more info.}

\item{types}{(optional): vector of resource type}

\item{size}{(optional): number of records to return (default = 10)}

\item{fetchAll}{(optional): Boolean whether to return all results for the query}

\item{fields}{(optional): vector specifying which fields to return. Returns all by default}

\item{sort}{(optional): field to sort by.  Add `-` to sort in descending order}

\item{facets}{(optional): field by which to aggregate (count) their frequency}

\item{facet_size}{(optional): how many facet groups to include in the facet total (default = 10)}
}
\value{

}
\description{
Title
}
\examples{
# Get the date latest date for every resource (newest of dateModified, dateCreated, datePublished)
resources_by_date = getResourcesData(query = "date:[2020-01-01 TO *]", fields = "date")
}
