% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResourcesResponse.R
\name{getResourcesResponse}
\alias{getResourcesResponse}
\title{Title}
\usage{
getResourcesResponse(
  queryString = NULL,
  size = 10,
  fetchAll = FALSE,
  fields = NULL,
  sort = NULL,
  facets = NULL,
  facet_size = 10
)
}
\arguments{
\item{queryString}{(optional): constructs a query over ALL fields, or a fielded query searching within specific fields. Fielded query terms should be separated by ` AND ` or ` OR `}

\item{size}{(optional): number of records to return (default = 10)}

\item{fetchAll}{(optional): Boolean whether to return all results for the query}

\item{fields}{(optional): vector specifying which fields to return. Returns all by default}

\item{sort}{(optional): field to sort by.  Add `-` to sort in descending order}

\item{facets}{(optional): field by which to aggregate (count) their frequency}

\item{facet_size}{(optional): how many facet groups to include in the facet total (default = 10)}
}
\value{
dataframe containing the response of the API query
}
\description{
Title
}
\examples{
# Return only the first 10 hits for 
df = getResourcesResponse("ivermectin", size = 10)
df = getResourcesResponse("_exists_:topicCategory", size = 10)
df = getResourcesResponse("ivermectin", size = 10, facets="@type", fields=c("date", "name"))
}
