% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupSublineages.R
\name{lookupSublineages}
\alias{lookupSublineages}
\title{Lookup sublineages for a given Pango lineage or WHO name}
\usage{
lookupSublineages(lineage, returnQueryString = FALSE)
}
\arguments{
\item{lineage}{String with the name of a Pango lineage or WHO name}

\item{returnQueryString}{Boolean to return a query string to be piped into functions like \link[outbreakinfo]{getPrevalence} (collapses vector by `"OR"`)}
}
\value{
list containing the sublineages associated with that lineage or WHO designation
}
\description{
Retrieve all sublineages of a given Pango lineage or WHO designation.
Based on a phylogeny maintained by the [Pango team](https://github.com/cov-lineages/lineages-website/blob/master/data/lineages.yml)
to grab all sublineages associated with a lineage, and a curated list of Variants of Concern, Variants of Interest,
Variants Under Monitoring, and de-escalated variants maintained by the [outbreak.info team](https://outbreak.info/situation-reports)
}
\examples{
# WHO-designated lineages
lookupSublineages("Delta")
lookupSublineages("epsilon")
lookupSublineages("epsilon", returnQueryString = TRUE)
# Pango lineage
lookupSublineages("B.1.1.7")
#' # Not a recognized lineage
lookupSublineages("VOC-21APR-02")
}
