% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticateUser.R
\docType{data}
\name{OUTBREAK_INFO_AUTH}
\alias{OUTBREAK_INFO_AUTH}
\title{Authenticate API}
\format{
An object of class \code{character} of length 1.
}
\usage{
OUTBREAK_INFO_AUTH
}
\value{
url
}
\description{
Authenticate API to get access to data
}
\examples{
authenticateUser()

}
\keyword{datasets}
