% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEpiData.R
\name{plotEpiData}
\alias{plotEpiData}
\title{Plot COVID-19 data of interest}
\usage{
plotEpiData(locations, variable)
}
\arguments{
\item{locations}{a vector or list of location names}

\item{variable}{metric to plot}
}
\value{
ggplot2 object
}
\description{
Plot a metric of interest using up-to-date COVID-19 data from outbreak.info for location(s) of interest (World Bank region, country, state/province, metropolitan area, county).
}
\examples{
p = plotEpiData("Florida", "confirmed_per_100k")
show(p)

}
