% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutationHeatmap.R
\name{plotMutationHeatmap}
\alias{plotMutationHeatmap}
\title{Title}
\usage{
plotMutationHeatmap(df, genes = NULL, title = NULL, lightBorders = TRUE)
}
\arguments{
\item{df:}{Resulting dataframe from calling \link[outbreakinfo]{getMutationsByLineage}}

\item{genes:}{(optional) vector containing which genes to include, e.g. c("Orf1a", "Orf1b", "S")}

\item{title:}{(optional) title to add to the plot}

\item{lightBorders:}{(optional) boolean; whether the borders between grid items should be separated with a light or dark border.}
}
\value{

}
\description{
Plots a heatamp of mutation prevalence across particular lineages.
}
\examples{
p1 = getMutationsByLineage(pangolin_lineage = "P.1", frequency = 0.5)
plotMutationHeatmap(p1, genes =c("Orf1a", "Orf1b", "S"))

delta_lineages = lookupSublineages("Delta", returnQueryString = FALSE)
delta_mutations = getMutationsByLineage(pangolin_lineage=delta_lineages, frequency=0.75)
plotMutationHeatmap(delta_mutations, genes = c("S"), title = "S-gene mutations in Delta lineages")
}
