% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrevalenceByLocation.R
\name{plotPrevalenceByLocation}
\alias{plotPrevalenceByLocation}
\title{Plot daily prevalence of a lineage by location}
\usage{
plotPrevalenceByLocation(
  pangolin_lineage,
  location,
  mutations = NULL,
  cumulative = NULL,
  include_title = F
)
}
\arguments{
\item{pangolin_lineage:}{(optional) PANGO lineage name or vector of PANGO lineage names. Either `pangolin_lineage` or `mutations` needs to be specified. A list of lineages will return a long dataframe with `lineage` as a variable; if you want to calculate the prevalence of lineage1 or lineage2 together, enter the lineages separated by " OR ". For instance, to calculate the prevalence of Delta, you'll need to supply `"B.1.617.2 OR AY.1 OR AY.2 OR ..."` **Be sure to include the space around "OR" and it must be capitalized.**}

\item{mutations:}{(optional) a `vector` of mutation(s). Either `pangolin_lineage` or `mutations` needs to be specified.}

\item{location:}{(optional) a location name}

\item{cumulative:}{(optional) `Boolean` (T/F), T returns cumulative prevalence since first day of detection}

\item{include_title:}{`Boolean` (T/F), T returns plot with title, F returns plot without title (default=F)}
}
\value{
ggplot2 object
}
\description{
Plots the daily prevalence of a PANGO lineage by location
}
\examples{
plotPrevalenceByLocation(pangolin_lineage = "P.1", location = "Brazil")


}
