% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrevalenceDotPlot.R
\name{plotPrevalenceDotPlot}
\alias{plotPrevalenceDotPlot}
\title{Plots a dot plot of prevalences of a variant in a specific location}
\usage{
plotPrevalenceDotPlot(df, title = NULL, subtitle = NULL, dot_size = 3)
}
\arguments{
\item{df}{Plots results from calling \link[outbreakinfo]{getCumulativeBySubadmin}}

\item{title}{(optional)}

\item{subtitle}{(optional)}
}
\value{

}
\description{
Plots a dot plot of prevalences of a variant in a specific location
}
\examples{
delta_by_state = getCumulativeBySubadmin(pangolin_lineage = delta_lineages, location = locationName, ndays = 90)
plotPrevalenceDotPlot(delta_by_state, "Delta prevalence over the last 90 days in the United States", subtitle = paste0("As of ", format(Sys.Date(), '\%d \%B \%Y')))
}
