% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrevalenceOverTime.R
\name{plotPrevalenceOverTime}
\alias{plotPrevalenceOverTime}
\title{Plot daily prevalence of a lineage by location}
\usage{
plotPrevalenceOverTime(
  df,
  colorVar = "lineage",
  title = "Prevalence over time",
  labelDictionary = NULL
)
}
\arguments{
\item{df}{result of the call to \link[outbreakinfo]{getPrevalence}}

\item{colorVar}{variable to used to color the line traces. `lineage` by default}

\item{title}{(optional) Title to add to the plot}

\item{labelDictionary}{(optional) a named list of values to replace in the plot legend.}
}
\value{
ggplot2 object
}
\description{
Plots the daily prevalence of a PANGO lineage by location
}
\examples{
p1_brazil <- getPrevalence(pangolin_lineage = "P.1", location = "Brazil")
plotPrevalenceOverTime(p1_brazil)

us <- getPrevalence(pangolin_lineage = c("B.1.1.7", "B.1.427 OR B.1.429", "B.1.617.2"), location = "United States")
plotPrevalenceOverTime(us, labelDictionary = c("B.1.427 OR B.1.429" = "Epsilon"))

# Overlay locations
b117_india = getPrevalence(pangolin_lineage = "B.1.1.7", location = "India")
b117_us = getPrevalence(pangolin_lineage = "B.1.1.7", location = "United States")
b117_uk = getPrevalence(pangolin_lineage = "B.1.1.7", location = "United Kingdom")
b117 = dplyr::bind_rows(b117_uk, b117_india, b117_us)
plotPrevalenceOverTime(b117, colorVar = "location", title="B.1.1.7 prevalence over time")
}
