% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchLocations.R
\name{searchLocations}
\alias{searchLocations}
\title{Get exact location names}
\usage{
searchLocations(locations_to_search, admin_level)
}
\arguments{
\item{locations_to_search}{vector or list of location (World Bank region, country, state/province, metropolitan area, county) name(s) at the same administrative level}

\item{admin_level}{an integer representing an administrative level (World Bank regions = -1, countries = 0, states/provinces = 1, metropolitan areas = 1.5, counties = 2)}
}
\value{
a vector or list of location names
}
\description{
Get exact spelling of locations at the same administrative level.
}
\examples{
searchLocations(c("California", "Florida", "Texas"), admin_level=1)

}
