% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAuthToken.R
\name{setAuthToken}
\alias{setAuthToken}
\title{Set authentication token}
\usage{
setAuthToken(token)
}
\arguments{
\item{token}{A String representing the token to be stored}
}
\description{
Writes authentication token to file
}
