% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalenceByLocation.R
\name{getPrevalenceByLocation}
\alias{getPrevalenceByLocation}
\title{Retrieve daily prevalence of a lineage by location}
\usage{
getPrevalenceByLocation(
  pangolin_lineage,
  location,
  mutations = NULL,
  cumulative = NULL
)
}
\arguments{
\item{pangolin_lineage:}{PANGO lineage name}

\item{location:}{a location name}

\item{mutations:}{(optional) a `vector` of mutation(s)}

\item{cumulative:}{(optional) `Boolean` (T/F), T returns cumulative prevalence since first day of detection}
}
\value{
dataframe
}
\description{
Retrieves the daily prevalence of a PANGO lineage by location
}
\examples{
getPrevalenceByLocation(pangolin_lineage = "P.1", location = "Brazil")

}
