% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubmissionDateByLocation.R
\name{getSubmissionDateByLocation}
\alias{getSubmissionDateByLocation}
\title{Retrieve most recent submission date}
\usage{
getSubmissionDateByLocation(
  pangolin_lineage,
  location = NULL,
  mutations = NULL
)
}
\arguments{
\item{pangolin_lineage}{PANGO lineage name}

\item{location}{(optional) a location name (if not specified, submission date at the global level returned)}

\item{mutations}{(optional) a `vector` of mutation(s)}
}
\value{
dataframe
}
\description{
Retrieves number of sequences by location, by date of sequence submission. 
See \link[outbreakinfo]{getSeqCounts} to view similar information by date of sample collection.
}
\examples{
getSubmissionDateByLocation(pangolin_lineage="B.1.1.7", location="California")

}
