% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutationHeatmap.R
\name{plotMutationHeatmap}
\alias{plotMutationHeatmap}
\title{Title Compare mutation prevalences between lineages}
\usage{
plotMutationHeatmap(df, gene2Plot = "S", title = NULL, lightBorders = TRUE)
}
\arguments{
\item{df}{Resulting dataframe from calling \link[outbreakinfo]{getMutationsByLineage}}

\item{gene2Plot}{(optional) string containing which genes to include, e.g. "Orf1a" By default, will limit the mutations to those in the S-gene.}

\item{title}{(optional) title to add to the plot}

\item{lightBorders}{(optional) boolean; whether the borders between grid items should be separated with a light or dark border.}
}
\value{

}
\description{
Plots a heatamp of mutation prevalence across particular lineages.
}
\examples{
p1 = getMutationsByLineage(pangolin_lineage = "P.1", frequency = 0.5)
plotMutationHeatmap(p1, gene2Plot = "ORF1a")

delta_lineages = lookupSublineages("Delta", returnQueryString = FALSE)
delta_mutations = getMutationsByLineage(pangolin_lineage=delta_lineages, frequency=0.75)
plotMutationHeatmap(delta_mutations, title = "S-gene mutations in Delta lineages")
}
