% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrevalenceDotPlot.R
\name{plotPrevalenceDotPlot}
\alias{plotPrevalenceDotPlot}
\title{Plots a dot plot of prevalences of a variant in a specific location}
\usage{
plotPrevalenceDotPlot(df, title = NULL, subtitle = NULL, dot_size = 3)
}
\arguments{
\item{df}{Dataframe resulting from calling \link[outbreakinfo]{getCumulativeBySubadmin}}

\item{title}{(optional) Title to include on the plot}

\item{subtitle}{(optional) Subtitle to include on the plot}

\item{dot_size}{(optional) Dot plot size (in mm)}
}
\value{

}
\description{
Plots a dot plot of prevalences of a variant in a specific location
}
\examples{

# Calculate the proportion of sequences assigned to BA.2 or BA.2.12.1 by U.S. State
prev_by_state = getCumulativeBySubadmin(pangolin_lineage = c("BA.2", "BA.2.12.1"), location = "United States", ndays = 90)

# Plot the results
plotPrevalenceDotPlot(prev_by_state, "Prevalence over the last 90 days in the United States", subtitle = paste0("As of ", format(Sys.Date(), '\%d \%B \%Y')))
}
